+++ src/src.mk
@@ -1,3 +1,4 @@
 
 PROJSRC  = src/main.c
 PROJSRC += src/blocking_uart_driver.c
+PROJSRC += src/motor_pwm.c
